+++ src/tests/itemsynctests.h
@@ -47,9 +47,11 @@ private slots:
     void addItemsWhenFull();
 
     void addItemsWhenFullOmitDeletingNotOwned();
+    void prioritizeNewFilesWhenItemLimitIsReached();
 
     void moveOwnItemsSortsBaseNames();
     void moveOwnItemsKeepsLargeTextData();
+    void moveLargeItemsToNonSyncTabKeepsData();
 
     void avoidDuplicateItemsAddedFromClipboard();
 
@@ -58,6 +60,7 @@ private slots:
     void sortItems();
 
     void copyFiles();
+    void copyFilesFromUriListIsAllOrNothing();
 
     void encryptionShouldNotAffectFiles();